+++ Phoenix/src/Window.cpp
@@ -146,7 +146,7 @@ namespace Phoenix{
 
 	void Window::fullscreen(){
 		glfwSetWindowAttrib(_window, GLFW_DECORATED, GLFW_FALSE);
-		glfwSetWindowPos(_window, 0, 0);
+		// glfwSetWindowPos(_window, 0, 0);
 		glfwSetWindowSize(_window, _config.width, _config.height);
 	}